feat: 챌린지 생성 / 챌린지 정보 조회 / 챌린지 포기 API Test 완료(Controller, Service 테스트 커버리지 100%) #15

+++ src/main/java/com/ceos/bankids/service/ChallengeServiceImpl.java
@@ -4,6 +4,7 @@ import com.ceos.bankids.controller.request.ChallengeRequest;
 import com.ceos.bankids.domain.Challenge;
 import com.ceos.bankids.domain.ChallengeCategory;
 import com.ceos.bankids.domain.ChallengeUser;
+import com.ceos.bankids.domain.Progress;
 import com.ceos.bankids.domain.TargetItem;
 import com.ceos.bankids.domain.User;
 import com.ceos.bankids.dto.ChallengeDTO;
@@ -13,7 +14,9 @@ import com.ceos.bankids.exception.NotFoundException;
 import com.ceos.bankids.repository.ChallengeCategoryRepository;
 import com.ceos.bankids.repository.ChallengeRepository;
 import com.ceos.bankids.repository.ChallengeUserRepository;
+import com.ceos.bankids.repository.ProgressRepository;
 import com.ceos.bankids.repository.TargetItemRepository;
+import java.util.List;
 import java.util.Objects;
 import java.util.Optional;
 import lombok.RequiredArgsConstructor;
@@ -31,6 +34,7 @@ public class ChallengeServiceImpl implements ChallengeService {
     private final ChallengeCategoryRepository challengeCategoryRepository;
     private final TargetItemRepository targetItemRepository;
     private final ChallengeUserRepository challengeUserRepository;
+    private final ProgressRepository progressRepository;
 
     @Transactional
     @Override
@@ -90,11 +94,18 @@ public class ChallengeServiceImpl implements ChallengeService {
         if (deleteChallengeUserRow.isPresent()) {
             ChallengeUser deleteChallengeUser = deleteChallengeUserRow.get();
             Challenge deleteChallenge = deleteChallengeUser.getChallenge();
+            Long deleteChallengeId = deleteChallenge.getId();
             if (!Objects.equals(deleteChallengeUser.getUser().getId(), user.getId())) {
                 throw new ForbiddenException("권한이 없습니다.");
             }
-            challengeUserRepository.delete(deleteChallengeUser);
-            challengeRepository.delete(deleteChallenge);
+            List<Progress> progressList = deleteChallenge.getProgressList();
+            if ((long) progressList.size() == 1 || progressList.isEmpty()) {
+                challengeUserRepository.delete(deleteChallengeUser);
+                challengeRepository.delete(deleteChallenge);
+            } else {
+                throw new BadRequestException("생성한지 일주일이 지난 돈길은 포기가 불가능합니다.");
+            }
+
             return null;
         } else {
             throw new NotFoundException("챌린지가 없습니다.");
